refactor(mcp): move buildScoreSubjectUrl into product-url (LFE-11051)
