+++ web/src/features/mcp/features/scores/tools/getScore.ts
@@ -3,7 +3,7 @@ import { z } from "zod";
 import { defineTool } from "../../../core/define-tool";
 import { runMcpTool } from "../../../core/run-mcp-tool";
 import { listScoresV3ForPublicApi } from "@/src/features/public-api/server/scores-api-v3";
-import { buildScoreSubjectUrl } from "../lib/subject-url";
+import { buildScoreSubjectUrl } from "@/src/utils/product-url";
 
 const GetScoreInputSchema = z.object({ scoreId: z.string() }).strict();
